feat: TextMate grammar for jikespg grammars
jikespg is the parser generator used in ECJ.
This TextMate grammar provides basic syntax highlighting for the grammar
files that jikespg uses.
In order to try out this grammar,
try editing
https://github.com/eclipse-jdt/eclipse.jdt.core/blob/master/org.eclipse.jdt.core.compiler.batch/grammar/java.g
,
or try editing some of the example grammars in the jikespg repo:
https://github.com/daveshields/jikespg

I wrote this grammar.
It doesn't come from the VS Code repository,
so it can't be auto-updated from there.

Signed-off-by: David Thompson <davthomp@redhat.com>